Keep related item view types webview safe
Move the resolved related-item shape into a view-safe type module so webview message types do not import service code and its Node-only transitive dependencies.
